fix(app, service, input, workflow, generate initialize): bug where the output flag is not applied in the init command
There was a bug where the init command did not apply the output flag.
As a result, the generated files always used the default output format.
