--- internal/generator/workdirprovider/workdir_provider.go
@@ -1,13 +1,0 @@
-package workdirprovider
-
-import "os"
-
-type WorkdirProvider struct{}
-
-func NewWorkdirProvider() *WorkdirProvider {
-	return &WorkdirProvider{}
-}
-
-func (w *WorkdirProvider) GetConfigOutputPath() (string, error) {
-	return os.Getwd()
-}
