+++ internal/workflow/initialize/initialize_template_json.go
@@ -2,42 +2,43 @@ package initialize
 
 import (
 	"fmt"
-	"path/filepath"
 
 	"github.com/taka1156/codespacegen/internal/domain/entity"
 	"github.com/taka1156/codespacegen/internal/domain/service"
+	"github.com/taka1156/codespacegen/internal/utils"
 )
 
 type InitializeSettingJson struct {
 	settingTemplateGenerator service.SettingTemplateGenerator
-	workdirProvider          service.WorkdirProvider
 	writer                   service.LocalFileWriter
 }
 
 func NewInitializeSettingJson(
 	settingTemplateGenerator service.SettingTemplateGenerator,
-	workdirProvider service.WorkdirProvider,
 	writer service.LocalFileWriter,
 ) *InitializeSettingJson {
 	return &InitializeSettingJson{
 		settingTemplateGenerator: settingTemplateGenerator,
-		workdirProvider:          workdirProvider,
 		writer:                   writer,
 	}
 }
 
-func (isj *InitializeSettingJson) Execute(templateJson entity.JsonConfig, settingJsonFileName string) error {
+func (isj *InitializeSettingJson) Execute(
+	templateJson entity.JsonConfig,
+	settingJsonFileName string,
+	outputDir string,
+) error {
 	file, err := isj.settingTemplateGenerator.Generate(templateJson)
 	if err != nil {
 		return fmt.Errorf("failed to generate template JSON: %w", err)
 	}
 
-	outputPath, err := isj.workdirProvider.GetConfigOutputPath()
+	outputPath, err := utils.ResolveOutputPath(outputDir, settingJsonFileName)
 	if err != nil {
-		return fmt.Errorf("failed to get config output path: %w", err)
+		return fmt.Errorf("failed to resolve output path: %w", err)
 	}
 
-	err = isj.writer.Write(filepath.Join(outputPath, settingJsonFileName), file, false)
+	err = isj.writer.Write(outputPath, file, false)
 	if err != nil {
 		return fmt.Errorf("failed to write template JSON: %w", err)
 	}
