+++ internal/workflow/initialize/initialize_template_json_test.go
@@ -17,14 +17,6 @@ func (m *mockSettingTemplateGenerator) Generate(tj entity.JsonConfig) (string, e
 	return m.generateFunc(tj)
 }
 
-type mockWorkdirProvider struct {
-	getConfigOutputPathFunc func() (string, error)
-}
-
-func (m *mockWorkdirProvider) GetConfigOutputPath() (string, error) {
-	return m.getConfigOutputPathFunc()
-}
-
 type mockLocalFileWriter struct {
 	writeFunc func(path string, content string, overwrite bool) error
 }
@@ -39,11 +31,6 @@ func TestInitializeSettingJson_Execute_Success(t *testing.T) {
 			return "test-content", nil
 		},
 	}
-	wd := &mockWorkdirProvider{
-		getConfigOutputPathFunc: func() (string, error) {
-			return "/tmp/", nil
-		},
-	}
 	writer := &mockLocalFileWriter{
 		writeFunc: func(path string, content string, overwrite bool) error {
 			if path != "/tmp/setting.json" || content != "test-content" || overwrite != false {
@@ -52,8 +39,8 @@ func TestInitializeSettingJson_Execute_Success(t *testing.T) {
 			return nil
 		},
 	}
-	isj := NewInitializeSettingJson(gen, wd, writer)
-	err := isj.Execute(entity.JsonConfig{}, "setting.json")
+	isj := NewInitializeSettingJson(gen, writer)
+	err := isj.Execute(entity.JsonConfig{}, "setting.json", "/tmp")
 	assert.NoError(t, err)
 }
 
@@ -63,27 +50,25 @@ func TestInitializeSettingJson_Execute_GenerateError(t *testing.T) {
 			return "", errors.New("generate error")
 		},
 	}
-	wd := &mockWorkdirProvider{getConfigOutputPathFunc: func() (string, error) { return "/", nil }}
 	writer := &mockLocalFileWriter{writeFunc: func(string, string, bool) error { return nil }}
-	isj := NewInitializeSettingJson(gen, wd, writer)
-	err := isj.Execute(entity.JsonConfig{}, "setting.json")
+	isj := NewInitializeSettingJson(gen, writer)
+	err := isj.Execute(entity.JsonConfig{}, "setting.json", "/tmp")
 	assert.ErrorContains(t, err, "failed to generate template JSON")
 }
 
-func TestInitializeSettingJson_Execute_GetConfigOutputPathError(t *testing.T) {
+func TestInitializeSettingJson_Execute_ResolveOutputPathError(t *testing.T) {
 	gen := &mockSettingTemplateGenerator{generateFunc: func(entity.JsonConfig) (string, error) { return "ok", nil }}
-	wd := &mockWorkdirProvider{getConfigOutputPathFunc: func() (string, error) { return "", errors.New("path error") }}
 	writer := &mockLocalFileWriter{writeFunc: func(string, string, bool) error { return nil }}
-	isj := NewInitializeSettingJson(gen, wd, writer)
-	err := isj.Execute(entity.JsonConfig{}, "setting.json")
-	assert.ErrorContains(t, err, "failed to get config output path")
+	isj := NewInitializeSettingJson(gen, writer)
+	// "../escaping.json" はパス外に出るためエラーになる
+	err := isj.Execute(entity.JsonConfig{}, "../escaping.json", "/tmp")
+	assert.ErrorContains(t, err, "failed to resolve output path")
 }
 
 func TestInitializeSettingJson_Execute_WriteError(t *testing.T) {
 	gen := &mockSettingTemplateGenerator{generateFunc: func(entity.JsonConfig) (string, error) { return "ok", nil }}
-	wd := &mockWorkdirProvider{getConfigOutputPathFunc: func() (string, error) { return "/tmp/", nil }}
 	writer := &mockLocalFileWriter{writeFunc: func(string, string, bool) error { return errors.New("write error") }}
-	isj := NewInitializeSettingJson(gen, wd, writer)
-	err := isj.Execute(entity.JsonConfig{}, "setting.json")
+	isj := NewInitializeSettingJson(gen, writer)
+	err := isj.Execute(entity.JsonConfig{}, "setting.json", "/tmp")
 	assert.ErrorContains(t, err, "failed to write template JSON")
 }
